requirements.txt
django
pillow
google-generativeai
django-crispy-forms
crispy-bootstrap5
python-dotenv
